Refactored main
